Implement intellisense fix
